Iter18: tokengenerator moved to separated folder. added some doc with example for this.
